File structure changes,
also tried to fix merging error on Git
