+++ src/main/java/UI/C_NewOrLoadScreen.java
@@ -1,4 +1,4 @@
-package gameUI;
+package UI;
 
 public class C_NewOrLoadScreen {
 }
