+++ src/main/java/UI/D_LoadGameScreen.java
@@ -1,4 +1,4 @@
-package gameUI;
+package UI;
 
 public class D_LoadGameScreen {
 }
