+++ src/main/java/UI/D_NewGameScreen.java
@@ -1,4 +1,4 @@
-package gameUI;
+package UI;
 
 public class D_NewGameScreen {
 }
