+++ src/main/java/UI/E_CheckProgressScreen.java
@@ -1,4 +1,4 @@
-package gameUI;
+package UI;
 
 public class E_CheckProgressScreen {
 }
