+++ src/main/java/UI/E_LeaderboardScreen.java
@@ -1,4 +1,4 @@
-package gameUI;
+package UI;
 
 public class E_LeaderboardScreen {
 }
