+++ src/main/java/UI/E_ShareScreen.java
@@ -1,4 +1,4 @@
-package gameUI;
+package UI;
 
 public class E_ShareScreen {
 }
