+++ src/main/java/game_UI/A_MainDisplay.java
@@ -1,0 +1,22 @@
+package game_UI;
+
+import java.awt.event.ActionListener;
+
+public class A_MainDisplay {
+
+    //Class Instantiation
+
+    public static void main(String[] args) {
+        new A_MainDisplay();
+    }
+
+    public A_MainDisplay(){
+        B_TitleScreen.createTitleScreen();
+    }
+
+    public class ChoiceHandler implements ActionListener {
+        public void actionPerformed(ActionEvent event){
+            String 
+        }
+    }
+}
