+++ src/main/java/game_UI/B_TitleScreen.java
@@ -1,0 +1,4 @@
+package game_UI;
+
+public class B_TitleScreen {
+}
