+++ src/main/java/game_UI/D_NewOrLoadScreen.java
@@ -1,0 +1,4 @@
+package game_UI;
+
+public class D_NewOrLoadScreen {
+}
